Tweaked how mods are added to towers
This may need further updating as more complexity is added to this system.
Currently the mod object keeps track of all objects it has not applied a mod to and then applies the mod to those towers.
Functionality needs to be added for removing mods when the mod object is destroyed.